pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-butler"
dynamic = ["version"]
description = ""
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []

authors = [
    { name = "Florent Clarret", email = "florent.clarret@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "click==8.1.7",
    "platformdirs==4.2.0",
    "pygithub==2.3.0",
    "pydantic==2.6.4",
    "gitpython==3.1.42",
    "rich==13.7.1",
    "tomli==2.0.1",
    "tomli-w==1.0.0",
]

[project.urls]
Source = "https://github.com/FlorentClarret/code-butler"
History = "https://github.com/FlorentClarret/code-butler/blob/main/HISTORY.md"
Issues = "https://github.com/FlorentClarret/code-butler/issues"

[project.scripts]
code-butler = "code_butler.cli:code_butler"

[tool.hatch.version]
path = "src/code_butler/__about__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
]

[tool.hatch.build.targets.wheel]
packages = ["src/code_butler"]

[tool.mypy]
disallow_untyped_defs = false
disallow_incomplete_defs = false
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_no_return = false
warn_unused_ignores = true
